requirements.txt
alembic==1.17.0
bcrypt==5.0.0
black==25.9.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
uv==0.9.2
Werkzeug==3.1.3
WTForms==3.2.1
